test(sidepanel): split sidepanel test debt hotspots
