+++ tests/unit/sidepanel/sidepanel.test.js
@@ -1,146 +1,86 @@
 import { jest } from '@jest/globals';
 import fs from 'node:fs';
 import path from 'node:path';
-import { normalizeUrl, computeStableUrl } from '../../../scripts/utils/urlUtils.js';
-import { UI_MESSAGES } from '../../../scripts/config/shared/messages.js';
-import { RUNTIME_ACTIONS } from '../../../scripts/config/shared/runtimeActions.js';
-import { sanitizeApiError } from '../../../scripts/utils/ApiErrorSanitizer.js';
-import { sanitizeUrlForLogging } from '../../../scripts/utils/LogSanitizer.js';
-import Logger from '../../../scripts/utils/Logger.js';
 import {
-  SYNC_BUTTON_DEBOUNCE_MS,
-  OPEN_BUTTON_DEBOUNCE_MS,
-} from '../../../pages/sidepanel/sidepanelUI.js';
+  computeStableUrl,
+  createDeferred,
+  flushMicrotasks,
+  loadSidepanelForCurrentView,
+  Logger,
+  normalizeUrl,
+  UI_MESSAGES,
+} from './sidepanel.shared.js';
 
-// ---- Mocks ----
-jest.mock('../../../scripts/utils/urlUtils.js', () => ({
-  normalizeUrl: jest.fn(url => url),
-  computeStableUrl: jest.fn(),
-  isSafeStableUrl: jest.fn(url => {
-    try {
-      const parsedUrl = new URL(url);
-      return parsedUrl.protocol === 'http:' || parsedUrl.protocol === 'https:';
-    } catch {
-      return false;
-    }
-  }),
-  isRootUrl: jest.fn(url => {
-    try {
-      const u = new URL(url);
-      return (u.pathname === '/' || u.pathname === '') && u.search.length === 0;
-    } catch {
-      return false;
-    }
-  }),
-}));
+function setupStaleCurrentTabRace() {
+  const staleStableUrl = 'https://stale.example.com/article';
+  const freshStableUrl = 'https://fresh.example.com/article';
+  const deferredStableUrl = createDeferred();
 
-jest.mock('../../../scripts/utils/Logger.js', () => ({
-  __esModule: true,
-  default: {
-    success: jest.fn(),
-    start: jest.fn(),
-    ready: jest.fn(),
-    info: jest.fn(),
-    debug: jest.fn(),
-    warn: jest.fn(),
-    error: jest.fn(),
-  },
-}));
-
-function createDeferred() {
-  let resolve;
-  let reject;
-  const promise = new Promise((_resolve, _reject) => {
-    resolve = _resolve;
-    reject = _reject;
+  chrome.tabs.get.mockImplementation(async tabId => {
+    const tabUrls = {
+      701: 'https://stale.example.com/raw',
+      702: 'https://fresh.example.com/raw',
+    };
+    return { id: tabId, url: tabUrls[tabId] || 'https://fallback.example.com' };
   });
-  return { promise, resolve, reject };
+
+  chrome.tabs.sendMessage.mockImplementation(async tabId => {
+    if (tabId === 701) {
+      return deferredStableUrl.promise;
+    }
+    if (tabId === 702) {
+      return { stableUrl: freshStableUrl };
+    }
+    return { stableUrl: 'https://default.example.com' };
+  });
+
+  chrome.storage.local.get.mockImplementation(async key =>
+    resolveRaceStorageLookup(key, { freshStableUrl, staleStableUrl })
+  );
+
+  return { deferredStableUrl, freshStableUrl, staleStableUrl };
 }
 
-async function flushMicrotasks(times = 6) {
-  for (let i = 0; i < times; i += 1) {
-    await Promise.resolve();
+function resolveRaceStorageLookup(key, { freshStableUrl, staleStableUrl }) {
+  if (!Array.isArray(key)) {
+    return {};
   }
+  const requestedKeys = new Set(key);
+  if (requestedKeys.has(`page_${freshStableUrl}`)) {
+    return {
+      [`page_${freshStableUrl}`]: {
+        highlights: [{ id: 'fresh', text: 'fresh text', color: 'yellow' }],
+      },
+    };
+  }
+  if (requestedKeys.has(`page_${staleStableUrl}`)) {
+    return {
+      [`page_${staleStableUrl}`]: {
+        highlights: [{ id: 'stale', text: 'stale text', color: 'yellow' }],
+      },
+    };
+  }
+  return {};
 }
 
-// Chrome API polyfills
-globalThis.chrome = {
-  tabs: {
-    onActivated: { addListener: jest.fn() },
-    onUpdated: { addListener: jest.fn() },
-    query: jest.fn(),
-    get: jest.fn(),
-    create: jest.fn(),
-    sendMessage: jest.fn(),
-  },
-  storage: {
-    local: {
-      get: jest.fn(),
-      set: jest.fn(),
-      remove: jest.fn(),
-    },
-    onChanged: { addListener: jest.fn() },
-  },
-  runtime: {
-    sendMessage: jest.fn(),
-  },
-};
+async function runStaleThenFreshTabLoads(onActivated, deferredStableUrl, staleStableUrl) {
+  const staleLoad = onActivated({ tabId: 701 });
+  await Promise.resolve();
+  const freshLoad = onActivated({ tabId: 702 });
+  await freshLoad;
 
-describe('Sidepanel JS Logic', () => {
+  deferredStableUrl.resolve({ stableUrl: staleStableUrl });
+  await staleLoad;
+}
+
+function getLastArrayStorageLookupKeys() {
+  const keyedGetCalls = chrome.storage.local.get.mock.calls.filter(([arg]) => Array.isArray(arg));
+  return keyedGetCalls.at(-1)?.[0] || [];
+}
+
+describe('Sidepanel current view', () => {
   beforeEach(async () => {
-    jest.clearAllMocks();
-    jest.useFakeTimers();
-
-    document.body.innerHTML = `
-      <div id="loading-state" style="display:none">Loading...</div>
-      <div id="empty-state" style="display:none">
-        <p>Empty</p>
-        <div class="subtitle">Subtitle</div>
-      </div>
-      <div id="highlights-list" style="display:none"></div>
-      <aside id="unsaved-page-notice" class="unsaved-page-notice" role="status" hidden></aside>
-      <button id="start-highlight-button"></button>
-      <button id="sync-button"></button>
-      <button id="open-notion-button"></button>
-      <div id="status-message"></div>
-      <div id="unsynced-view" style="display:none"></div>
-      <div id="unsynced-toolbar" style="display:none">
-        <span id="unsynced-count-label"></span>
-        <button id="clear-all-btn"></button>
-      </div>
-      <button id="load-more-btn" style="display:none"></button>
-      <span id="unsynced-badge"></span>
-      <template id="highlight-card-template">
-        <div class="highlight-card">
-          <div class="highlight-color-indicator"></div>
-          <p class="highlight-text"></p>
-          <button class="delete-button"></button>
-        </div>
-      </template>
-      <template id="page-card-template">
-        <div class="page-card">
-          <div class="page-title"></div>
-          <div class="page-meta"></div>
-          <div class="page-card-previews"></div>
-          <div class="page-card-remaining"></div>
-          <button class="page-open-button"></button>
-          <button class="page-delete-button"></button>
-        </div>
-      </template>
-    `;
-
-    chrome.tabs.query.mockResolvedValue([{ id: 101, url: 'https://example.com' }]);
-    chrome.tabs.get.mockResolvedValue({ id: 102, url: 'https://example.org' });
-    chrome.tabs.create.mockResolvedValue({ id: 103, url: 'https://opened.example' });
-    chrome.tabs.sendMessage.mockResolvedValue({ stableUrl: 'https://example.js/stable' });
-    chrome.storage.local.get.mockResolvedValue({});
-
-    jest.isolateModules(() => {
-      require('../../../pages/sidepanel/sidepanel.js');
-    });
-
-    document.dispatchEvent(new Event('DOMContentLoaded'));
-    await flushMicrotasks();
+    await loadSidepanelForCurrentView();
   });
 
   afterEach(() => {
@@ -703,638 +643,6 @@ describe('Sidepanel JS Logic', () => {
     });
   });
 
-  describe('User Interactions', () => {
-    it('should delete highlight on click', async () => {
-      const currentMockData = {
-        'highlights_https://example.js/stable': {
-          highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-        },
-      };
-
-      chrome.storage.local.get.mockImplementation(async () => {
-        return currentMockData;
-      });
-
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-
-      chrome.storage.local.remove.mockResolvedValue();
-
-      const delBtn = document.querySelector('.delete-button');
-      delBtn.click();
-      // Phase 4：handleDelete shouldRemove 路徑 + helper-driven cleanup 增加多個 await，
-      // 需要更多 microtask 循環才能 drain remove + tabs.query + sendMessage。
-      for (let i = 0; i < 10; i++) {
-        await Promise.resolve();
-      }
-
-      expect(chrome.storage.local.remove).toHaveBeenCalled();
-      const removedKeys = chrome.storage.local.remove.mock.calls.flatMap(call =>
-        Array.isArray(call[0]) ? call[0] : [call[0]]
-      );
-      expect(removedKeys).toContain('highlights_https://example.js/stable');
-      expect(chrome.tabs.sendMessage).toHaveBeenCalledWith(101, {
-        action: 'REMOVE_HIGHLIGHT_DOM',
-        highlightId: '1',
-      });
-    });
-
-    it('should update storage if highlights remain after delete', async () => {
-      const currentMockData = {
-        'highlights_https://example.js/stable': {
-          highlights: [
-            { id: '1', text: 'hello', color: 'yellow' },
-            { id: '2', text: 'world', color: 'blue' },
-          ],
-          metadata: { lastUpdated: 1000, title: 'Example' },
-        },
-      };
-
-      chrome.storage.local.get.mockImplementation(async () => currentMockData);
-
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-      // Click delete the FIRST one
-      const delBtn = document.querySelector('.delete-button');
-      delBtn.click();
-      for (let i = 0; i < 5; i++) {
-        await Promise.resolve();
-      }
-
-      expect(chrome.storage.local.set).toHaveBeenCalled();
-      const args = chrome.storage.local.set.mock.calls[0][0];
-      expect(args['highlights_https://example.js/stable'].highlights).toHaveLength(1);
-      expect(args['highlights_https://example.js/stable']).not.toBe(
-        currentMockData['highlights_https://example.js/stable']
-      );
-      expect(args['highlights_https://example.js/stable'].metadata).not.toBe(
-        currentMockData['highlights_https://example.js/stable'].metadata
-      );
-      expect(currentMockData['highlights_https://example.js/stable'].highlights).toHaveLength(2);
-      expect(currentMockData['highlights_https://example.js/stable'].metadata).toEqual({
-        lastUpdated: 1000,
-        title: 'Example',
-      });
-    });
-
-    it('在新的 page_* 格式上應能正確處理 handleDelete（部分刪除與完全移除）', async () => {
-      // test remove full
-      let fakeStore = {
-        'page_https://example.js/stable': {
-          highlights: [{ id: '1', text: 'h1', color: 'yellow' }],
-        },
-      };
-      chrome.storage.local.get.mockImplementation(async k => {
-        if (typeof k === 'string') {
-          return { [k]: fakeStore[k] };
-        }
-        if (Array.isArray(k)) {
-          const result = {};
-          for (const key of k) {
-            if (fakeStore[key]) {
-              result[key] = fakeStore[key];
-            }
-          }
-          return result;
-        }
-        return fakeStore;
-      });
-
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-      await Promise.resolve();
-
-      let delBtn = document.querySelector('.delete-button');
-      if (delBtn) {
-        delBtn.click();
-      }
-      // Phase 4：handleDelete 的 shouldRemove 路徑增加 _resolvePageDeletionKeys 內部的多次 await，
-      // 需要更多 microtask 循環才能 drain 所有 storage operation。
-      for (let i = 0; i < 6; i++) {
-        await Promise.resolve();
-      }
-      const removedKeys1 = chrome.storage.local.remove.mock.calls.flatMap(call =>
-        Array.isArray(call[0]) ? call[0] : [call[0]]
-      );
-      expect(removedKeys1).toContain('page_https://example.js/stable');
-
-      // test partial
-      chrome.storage.local.remove.mockClear();
-      chrome.storage.local.set.mockClear();
-      fakeStore = {
-        'page_https://example.js/stable': {
-          highlights: [
-            { id: '1', text: 'h1', color: 'yellow' },
-            { id: '2', text: 'h2', color: 'blue' },
-          ],
-        },
-      };
-
-      await onActivated({ tabId: 600 });
-      await Promise.resolve();
-      delBtn = document.querySelector('.delete-button');
-      if (delBtn) {
-        delBtn.click();
-      }
-      await Promise.resolve();
-      await Promise.resolve();
-      expect(chrome.storage.local.set).toHaveBeenCalled();
-    });
-
-    it('在舊的 array 格式上應能正確處理 handleDelete', async () => {
-      // test remove full
-      let fakeStore = {
-        'highlights_https://example.js/stable': {
-          highlights: [{ id: '1', text: 'h1', color: 'red' }],
-        },
-      };
-      // For this legacy array format, when handleDelete reads, we want it to read array, but when render Highlights reads, we want it to have { highlights: [] } so it renders.
-      // Wait, renderHighlights also accepts array if properly formatted! Let's check:
-      chrome.storage.local.get.mockImplementation(async k => {
-        // If it's a direct pull by handleDelete, pretend it's an array to trigger legacy code path
-        if (typeof k === 'string') {
-          return { [k]: [{ id: '1', text: 'h1', color: 'red' }] };
-        }
-        if (Array.isArray(k)) {
-          const result = {};
-          for (const key of k) {
-            if (fakeStore[key]) {
-              result[key] = fakeStore[key];
-            }
-          }
-          return result;
-        }
-        return fakeStore;
-      });
-
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-      await Promise.resolve();
-
-      let delBtn = document.querySelector('.delete-button');
-      if (delBtn) {
-        delBtn.click();
-      }
-      for (let i = 0; i < 6; i++) {
-        await Promise.resolve();
-      }
-      const removedKeysLegacy = chrome.storage.local.remove.mock.calls.flatMap(call =>
-        Array.isArray(call[0]) ? call[0] : [call[0]]
-      );
-      expect(removedKeysLegacy).toContain('highlights_https://example.js/stable');
-
-      // test partial
-      chrome.storage.local.remove.mockClear();
-      chrome.storage.local.set.mockClear();
-
-      fakeStore = {
-        'highlights_https://example.js/stable': {
-          highlights: [
-            { id: '1', text: 'h1', color: 'red' },
-            { id: '2', text: 'h2', color: 'blue' },
-          ],
-        },
-      };
-      chrome.storage.local.get.mockImplementation(async k => {
-        if (typeof k === 'string') {
-          return {
-            [k]: [
-              { id: '1', text: 'h1', color: 'red' },
-              { id: '2', text: 'h2', color: 'blue' },
-            ],
-          };
-        }
-        if (Array.isArray(k)) {
-          const result = {};
-          for (const key of k) {
-            if (fakeStore[key]) {
-              result[key] = fakeStore[key];
-            }
-          }
-          return result;
-        }
-        return fakeStore;
-      });
-
-      await onActivated({ tabId: 600 });
-      await Promise.resolve();
-      delBtn = document.querySelector('.delete-button');
-      if (delBtn) {
-        delBtn.click();
-      }
-      await Promise.resolve();
-      await Promise.resolve();
-      expect(chrome.storage.local.set).toHaveBeenCalled();
-    });
-
-    it('在 handleDelete 期間若遇到未知資料格式應能優雅地移除', async () => {
-      const fakeStore = {
-        'highlights_https://example.js/stable': {
-          highlights: [{ id: '1', text: 'h1', color: 'yellow' }],
-        },
-      };
-      chrome.storage.local.get.mockImplementation(async k => {
-        // give garbage on direct single-key get
-        if (typeof k === 'string') {
-          return { [k]: 'unknown_string_data' };
-        }
-        if (Array.isArray(k)) {
-          const result = {};
-          for (const key of k) {
-            if (fakeStore[key]) {
-              result[key] = fakeStore[key];
-            }
-          }
-          return result;
-        }
-        return fakeStore;
-      });
-
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-      await Promise.resolve();
-      const delBtn = document.querySelector('.delete-button');
-      if (delBtn) {
-        delBtn.click();
-      }
-      for (let i = 0; i < 6; i++) {
-        await Promise.resolve();
-      }
-      const removedKeysGarbage = chrome.storage.local.remove.mock.calls.flatMap(call =>
-        Array.isArray(call[0]) ? call[0] : [call[0]]
-      );
-      expect(removedKeysGarbage).toContain('highlights_https://example.js/stable');
-    });
-
-    it('should trigger sync click successfully', async () => {
-      chrome.storage.local.get.mockImplementation(async key => {
-        if (typeof key === 'string' && key.startsWith('saved_')) {
-          return { [key]: { notionPageId: 'page-123' } };
-        }
-        return {
-          'highlights_https://example.js/stable': {
-            highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-          },
-        };
-      });
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-
-      chrome.runtime.sendMessage.mockResolvedValue({ success: true, statusKind: 'saved' });
-
-      const syncBtn = document.querySelector('#sync-button');
-      syncBtn.click();
-      await Promise.resolve();
-      await Promise.resolve();
-
-      expect(chrome.runtime.sendMessage).toHaveBeenCalledWith({ action: 'savePage' });
-      expect(document.querySelector('#status-message').className).toBe('status-message success');
-
-      jest.runAllTimers();
-      expect(syncBtn.disabled).toBe(true);
-    });
-
-    it('[REGRESSION] sync button 不應在 storage 已改回 unsaved 後被 finally 無條件重新啟用', async () => {
-      const savedStore = {
-        'page_https://example.js/stable': {
-          highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-          notion: { pageId: 'page-123' },
-        },
-      };
-      const unsavedStore = {
-        'page_https://example.js/stable': {
-          highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-          notion: null,
-        },
-      };
-      let currentStore = savedStore;
-
-      chrome.storage.local.get.mockImplementation(async key => {
-        if (typeof key === 'string') {
-          return { [key]: currentStore[key] };
-        }
-        if (Array.isArray(key)) {
-          const result = {};
-          for (const item of key) {
-            if (item in currentStore) {
-              result[item] = currentStore[item];
-            }
-          }
-          return result;
-        }
-        return currentStore;
-      });
-
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      const onStorageChanged = chrome.storage.onChanged.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-      await flushMicrotasks();
-
-      chrome.runtime.sendMessage.mockResolvedValue({ success: true });
-
-      const syncBtn = document.querySelector('#sync-button');
-      expect(syncBtn.disabled).toBe(false);
-
-      syncBtn.click();
-      await flushMicrotasks();
-
-      currentStore = unsavedStore;
-      onStorageChanged(
-        {
-          'page_https://example.js/stable': {
-            oldValue: savedStore['page_https://example.js/stable'],
-            newValue: unsavedStore['page_https://example.js/stable'],
-          },
-        },
-        'local'
-      );
-      await flushMicrotasks();
-
-      jest.runAllTimers();
-      await flushMicrotasks();
-
-      expect(syncBtn.disabled).toBe(true);
-    });
-
-    it('should use named debounce constants when re-enabling open notion button', async () => {
-      chrome.storage.local.get.mockImplementation(async key => {
-        if (typeof key === 'string' && key.startsWith('saved_')) {
-          return { [key]: { notionPageId: 'page-123' } };
-        }
-        return {
-          'highlights_https://example.js/stable': {
-            highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-          },
-        };
-      });
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-      await Promise.resolve();
-      await Promise.resolve();
-
-      chrome.runtime.sendMessage
-        .mockResolvedValueOnce({ success: true, statusKind: 'saved' })
-        .mockResolvedValueOnce({ success: true });
-
-      const timeoutSpy = jest.spyOn(globalThis, 'setTimeout');
-
-      document.querySelector('#open-notion-button').click();
-      await Promise.resolve();
-      await Promise.resolve();
-
-      expect(timeoutSpy).toHaveBeenCalledWith(expect.any(Function), OPEN_BUTTON_DEBOUNCE_MS);
-
-      timeoutSpy.mockRestore();
-    });
-
-    it('should trigger sync click gracefully when fails', async () => {
-      chrome.storage.local.get.mockImplementation(async key => {
-        if (typeof key === 'string' && key.startsWith('saved_')) {
-          return { [key]: { notionPageId: 'page-123' } };
-        }
-        return {
-          'highlights_https://example.js/stable': {
-            highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-          },
-        };
-      });
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-
-      chrome.runtime.sendMessage.mockRejectedValue(new Error('Extension error message!'));
-
-      const syncBtn = document.querySelector('#sync-button');
-      syncBtn.click();
-      await Promise.resolve();
-      await Promise.resolve();
-      await Promise.resolve();
-
-      expect(document.querySelector('#status-message').className).toBe('status-message error');
-    });
-
-    it('should restore unsaved sync button state when save rejects after storage changes', async () => {
-      const sendMessage = createDeferred();
-      const savedStore = {
-        'page_https://example.js/stable': {
-          highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-          notion: { pageId: 'page-123' },
-        },
-      };
-      const unsavedStore = {
-        'page_https://example.js/stable': {
-          highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-          notion: null,
-        },
-      };
-      let currentStore = savedStore;
-
-      chrome.storage.local.get.mockImplementation(async key => {
-        if (typeof key === 'string') {
-          return { [key]: currentStore[key] };
-        }
-        if (Array.isArray(key)) {
-          const result = {};
-          for (const item of key) {
-            if (item in currentStore) {
-              result[item] = currentStore[item];
-            }
-          }
-          return result;
-        }
-        return currentStore;
-      });
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      const onStorageChanged = chrome.storage.onChanged.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-      await flushMicrotasks();
-
-      chrome.runtime.sendMessage.mockReturnValue(sendMessage.promise);
-
-      const syncBtn = document.querySelector('#sync-button');
-      expect(syncBtn.disabled).toBe(false);
-      expect(syncBtn.title).toBe('');
-
-      syncBtn.click();
-      await flushMicrotasks();
-
-      currentStore = unsavedStore;
-      onStorageChanged(
-        {
-          'page_https://example.js/stable': {
-            oldValue: savedStore['page_https://example.js/stable'],
-            newValue: unsavedStore['page_https://example.js/stable'],
-          },
-        },
-        'local'
-      );
-      await flushMicrotasks();
-
-      expect(syncBtn.disabled).toBe(true);
-      expect(syncBtn.title).toBe(UI_MESSAGES.SIDEPANEL.PAGE_NOT_SAVED);
-
-      sendMessage.reject(new Error('Extension error message!'));
-      await flushMicrotasks();
-
-      expect(syncBtn.disabled).toBe(true);
-      expect(syncBtn.title).toBe(UI_MESSAGES.SIDEPANEL.PAGE_NOT_SAVED);
-    });
-
-    it('should not display raw savePage error returned from runtime message', async () => {
-      chrome.storage.local.get.mockImplementation(async key => {
-        if (typeof key === 'string' && key.startsWith('saved_')) {
-          return { [key]: { notionPageId: 'page-123' } };
-        }
-        return {
-          'highlights_https://example.js/stable': {
-            highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-          },
-        };
-      });
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-      await flushMicrotasks();
-
-      chrome.runtime.sendMessage.mockResolvedValue({ success: false, error: 'Custom API Error' });
-
-      const syncBtn = document.querySelector('#sync-button');
-      syncBtn.click();
-      await Promise.resolve();
-      await Promise.resolve();
-      await Promise.resolve();
-
-      expect(document.querySelector('#status-message').textContent).toBe(
-        UI_MESSAGES.SIDEPANEL.SYNC_FAILED
-      );
-      expect(document.querySelector('#status-message').className).toBe('status-message error');
-      expect(Logger.error).toHaveBeenCalledWith(
-        '[SidePanel] savePage failed',
-        expect.objectContaining({
-          action: 'savePage',
-          result: 'failure',
-          error: expect.any(String),
-        })
-      );
-    });
-
-    it('should not display raw openNotionPage error returned from runtime message', async () => {
-      chrome.storage.local.get.mockImplementation(async key => {
-        if (typeof key === 'string' && key.startsWith('saved_')) {
-          return { [key]: { notionPageId: 'page-123' } };
-        }
-        return {
-          'highlights_https://example.js/stable': {
-            highlights: [{ id: '1', text: 'hello world', color: 'yellow' }],
-          },
-        };
-      });
-      const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-      await onActivated({ tabId: 600 });
-      await Promise.resolve();
-      await Promise.resolve();
-
-      chrome.runtime.sendMessage.mockResolvedValue({
-        success: false,
-        error: 'Leaked debug detail',
-      });
-
-      const openBtn = document.querySelector('#open-notion-button');
-      openBtn.click();
-      await Promise.resolve();
-      await Promise.resolve();
-      await Promise.resolve();
-
-      expect(document.querySelector('#status-message').textContent).toBe(
-        UI_MESSAGES.SIDEPANEL.OPEN_FAILED
-      );
-      expect(document.querySelector('#status-message').className).toBe('status-message error');
-      expect(Logger.error).toHaveBeenCalledWith(
-        '[SidePanel] openNotionPage failed',
-        expect.objectContaining({
-          error: expect.any(String),
-        })
-      );
-    });
-
-    it('應該成功觸發 startHighlight', async () => {
-      chrome.runtime.sendMessage.mockResolvedValue({ success: true });
-
-      const startBtn = document.querySelector('#start-highlight-button');
-      startBtn.click();
-      await Promise.resolve();
-      await Promise.resolve();
-
-      expect(chrome.runtime.sendMessage).toHaveBeenCalledWith({
-        action: RUNTIME_ACTIONS.START_HIGHLIGHT,
-      });
-      expect(document.querySelector('#status-message').className).toContain('success');
-    });
-
-    it('不應直接顯示 runtime message 回傳的原始 startHighlight 錯誤', async () => {
-      chrome.runtime.sendMessage.mockResolvedValue({
-        success: false,
-        error: 'Highlighter initialization failed',
-      });
-
-      const startBtn = document.querySelector('#start-highlight-button');
-      startBtn.click();
-      await Promise.resolve();
-      await Promise.resolve();
-
-      expect(document.querySelector('#status-message').textContent).not.toBe(
-        'Highlighter initialization failed'
-      );
-      expect(Logger.error).toHaveBeenCalledWith(
-        '[SidePanel] startHighlight failed',
-        expect.objectContaining({
-          action: 'startHighlight',
-          operation: 'highlight-init',
-          result: 'failure',
-          error: sanitizeApiError('Highlighter initialization failed', 'sidepanel_start_highlight'),
-        })
-      );
-    });
-
-    it('應該在 runtime sendMessage reject 時顯示格式化後的 startHighlight 錯誤', async () => {
-      const runtimeError = new Error('Network error');
-      chrome.runtime.sendMessage.mockRejectedValue(runtimeError);
-
-      const startBtn = document.querySelector('#start-highlight-button');
-      startBtn.click();
-      await Promise.resolve();
-      await Promise.resolve();
-
-      expect(document.querySelector('#status-message').textContent).toBe(
-        `${UI_MESSAGES.POPUP.HIGHLIGHT_FAILED_PREFIX}網路連線異常，請檢查網路後重試`
-      );
-      expect(Logger.error).toHaveBeenCalledWith(
-        '[SidePanel] startHighlight failed',
-        expect.objectContaining({
-          action: 'startHighlight',
-          operation: 'runtime-sendMessage',
-          result: 'failure',
-          error: runtimeError,
-          reason: sanitizeApiError(runtimeError, 'sidepanel_start_highlight'),
-        })
-      );
-    });
-
-    it('should re-enable start highlight button using named debounce constant', async () => {
-      chrome.runtime.sendMessage.mockRejectedValue(new Error('Extension error message!'));
-
-      const timeoutSpy = jest.spyOn(globalThis, 'setTimeout');
-      const startBtn = document.querySelector('#start-highlight-button');
-
-      startBtn.click();
-      await Promise.resolve();
-      await Promise.resolve();
-
-      expect(timeoutSpy).toHaveBeenCalledWith(expect.any(Function), SYNC_BUTTON_DEBOUNCE_MS);
-      expect(document.querySelector('#status-message').className).toContain('error');
-
-      timeoutSpy.mockRestore();
-    });
-  });
-
   describe('Storage Changes Sync', () => {
     it('should reload tab data if SC_HL key changes', async () => {
       const onStorageChanged = chrome.storage.onChanged.addListener.mock.calls[0][0];
@@ -1390,814 +698,20 @@ describe('Sidepanel JS Logic', () => {
     it('should not overwrite cached tab urls with stale loadCurrentTab results', async () => {
       const onActivated = chrome.tabs.onActivated.addListener.mock.calls[0][0];
       const onStorageChanged = chrome.storage.onChanged.addListener.mock.calls[0][0];
-      const staleStableUrl = 'https://stale.example.com/article';
-      const freshStableUrl = 'https://fresh.example.com/article';
-      const deferredStableUrl = createDeferred();
-
-      chrome.tabs.get.mockImplementation(async tabId => {
-        if (tabId === 701) {
-          return { id: 701, url: 'https://stale.example.com/raw' };
-        }
-        if (tabId === 702) {
-          return { id: 702, url: 'https://fresh.example.com/raw' };
-        }
-        return { id: tabId, url: 'https://fallback.example.com' };
-      });
-
-      chrome.tabs.sendMessage.mockImplementation(async tabId => {
-        if (tabId === 701) {
-          return deferredStableUrl.promise;
-        }
-        if (tabId === 702) {
-          return { stableUrl: freshStableUrl };
-        }
-        return { stableUrl: 'https://default.example.com' };
-      });
-
-      chrome.storage.local.get.mockImplementation(async key => {
-        if (Array.isArray(key)) {
-          const requestedKeys = new Set(key);
-          if (requestedKeys.has(`page_${freshStableUrl}`)) {
-            return {
-              [`page_${freshStableUrl}`]: {
-                highlights: [{ id: 'fresh', text: 'fresh text', color: 'yellow' }],
-              },
-            };
-          }
-          if (requestedKeys.has(`page_${staleStableUrl}`)) {
-            return {
-              [`page_${staleStableUrl}`]: {
-                highlights: [{ id: 'stale', text: 'stale text', color: 'yellow' }],
-              },
-            };
-          }
-          return {};
-        }
-        if (typeof key === 'string') {
-          return {};
-        }
-        return {};
-      });
+      const { deferredStableUrl, freshStableUrl, staleStableUrl } = setupStaleCurrentTabRace();
 
       chrome.storage.local.get.mockClear();
-
-      const staleLoad = onActivated({ tabId: 701 });
-      await Promise.resolve();
-      const freshLoad = onActivated({ tabId: 702 });
-      await freshLoad;
-
-      deferredStableUrl.resolve({ stableUrl: staleStableUrl });
-      await staleLoad;
+      await runStaleThenFreshTabLoads(onActivated, deferredStableUrl, staleStableUrl);
 
       chrome.storage.local.get.mockClear();
       onStorageChanged({ 'highlights_https://fresh.example.com/raw': { newValue: {} } }, 'local');
       await Promise.resolve();
       await Promise.resolve();
 
-      const keyedGetCalls = chrome.storage.local.get.mock.calls.filter(([arg]) =>
-        Array.isArray(arg)
-      );
-      const lastKeys = keyedGetCalls.at(-1)?.[0] || [];
+      const lastKeys = getLastArrayStorageLookupKeys();
 
       expect(lastKeys).toContain(`page_${freshStableUrl}`);
       expect(lastKeys).not.toContain(`page_${staleStableUrl}`);
     });
-  });
-}); // end describe('Sidepanel JS Logic')
-
-// ---- 共用 DOM helper（待同步視圖測試用） ----
-
-function buildUnsyncedDOM() {
-  document.body.innerHTML = `
-    <div id="loading-state" style="display:none"><div class="spinner"></div><p></p></div>
-    <div id="empty-state" style="display:none"><p>Empty</p><div class="subtitle"></div></div>
-    <div id="highlights-list" style="display:none"></div>
-    <button id="start-highlight-button"></button>
-    <div id="unsynced-view" style="display:none"></div>
-    <div id="unsynced-toolbar" style="display:none">
-      <span id="unsynced-count-label"></span>
-      <button id="clear-all-btn"></button>
-    </div>
-    <button id="load-more-btn" style="display:none">Load more</button>
-    <button id="sync-button"></button>
-    <button id="open-notion-button" style="display:none"></button>
-    <div id="status-message"></div>
-    <div class="view-tabs">
-      <button class="view-tab active" data-view="current">Current Page</button>
-      <button class="view-tab" data-view="unsynced">Pending<span id="unsynced-badge"></span></button>
-    </div>
-    <template id="highlight-card-template">
-      <div class="highlight-card">
-        <div class="highlight-color-indicator"></div>
-        <p class="highlight-text"></p>
-        <button class="delete-button"></button>
-      </div>
-    </template>
-    <template id="page-card-template">
-      <div class="page-card">
-        <div class="page-card-header">
-          <div class="page-title-row">
-            <span class="status-dot"></span>
-            <p class="page-title"></p>
-          </div>
-          <div class="page-info"><span class="page-meta"></span></div>
-          <button class="page-open-button"></button>
-          <button class="page-delete-button"></button>
-        </div>
-        <div class="page-card-previews"></div>
-        <span class="page-card-remaining"></span>
-      </div>
-    </template>
-  `;
-}
-
-async function initModule(storageMock) {
-  if (typeof storageMock === 'function') {
-    chrome.storage.local.get.mockImplementation(storageMock);
-  } else {
-    chrome.storage.local.get.mockResolvedValue(storageMock);
-  }
-  jest.isolateModules(() => {
-    require('../../../pages/sidepanel/sidepanel.js');
-  });
-  document.dispatchEvent(new Event('DOMContentLoaded'));
-  // 讓所有 async 完成（init 會平行跑 loadCurrentTab + updateUnsyncedBadge）
-  for (let i = 0; i < 10; i++) {
-    await Promise.resolve();
-  }
-}
-
-async function clickUnsyncedTab() {
-  const tab = document.querySelector('[data-view="unsynced"]');
-  tab.click();
-  for (let i = 0; i < 5; i++) {
-    await Promise.resolve();
-  }
-}
-
-describe('Unsynced View (getUnsyncedPages integration)', () => {
-  beforeEach(async () => {
-    jest.clearAllMocks();
-    jest.useFakeTimers();
-    buildUnsyncedDOM();
-
-    chrome.tabs.query.mockResolvedValue([{ id: 101, url: 'https://example.com' }]);
-    chrome.tabs.get.mockResolvedValue({ id: 102, url: 'https://example.org' });
-    chrome.tabs.create.mockResolvedValue({ id: 103, url: 'https://opened.example' });
-    chrome.tabs.sendMessage.mockResolvedValue({ stableUrl: 'https://example.com' });
-    chrome.storage.local.get.mockResolvedValue({});
-  });
-
-  afterEach(() => {
-    jest.useRealTimers();
-  });
-
-  it('should filter out synced pages (with notionPageId)', async () => {
-    await initModule({
-      'highlights_https://a.com/p': {
-        highlights: [{ id: '1', text: 'Synced text', color: 'yellow' }],
-        updatedAt: 2000,
-      },
-      'saved_https://a.com/p': { notionPageId: 'page-aaa' }, // 已同步
-      'highlights_https://b.com/p': {
-        highlights: [{ id: '2', text: 'Unsynced text', color: 'blue' }],
-        updatedAt: 1000,
-      },
-      // b.com 沒有 saved_，是未同步
-    });
-
-    // 切換到待同步 tab
-    await clickUnsyncedTab();
-
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    expect(cards).toHaveLength(1);
-    expect(cards[0].querySelector('.page-title').textContent).toContain('b.com');
-  });
-
-  it('[REGRESSION] should treat a non-object storage snapshot as an empty unsynced list', async () => {
-    await initModule(async key => {
-      if (key === null) {
-        return null;
-      }
-      return {};
-    });
-    Logger.error.mockClear();
-
-    await clickUnsyncedTab();
-
-    const emptyMessage = document.querySelector('#unsynced-view .unsynced-empty');
-    expect(emptyMessage.textContent).toBe(UI_MESSAGES.SIDEPANEL.ALL_SYNCED);
-    expect(Logger.error).not.toHaveBeenCalledWith(
-      '[SidePanel] renderUnsyncedView failed after tab switch',
-      expect.any(Object)
-    );
-  });
-
-  it('should show preview text truncated to 80 chars', async () => {
-    const longText = 'A'.repeat(100);
-    await initModule({
-      'highlights_https://c.com/p': {
-        highlights: [{ id: '1', text: longText, color: 'yellow' }],
-        updatedAt: 1000,
-      },
-    });
-
-    await clickUnsyncedTab();
-
-    const previewRow = document.querySelector('.preview-row');
-    expect(previewRow).not.toBeNull();
-    // 文字本身被截斷至 80 字元，加上省略號和引號後顯示
-    expect(previewRow.textContent).toContain('...');
-    expect(previewRow.textContent).toBe(`"${longText.slice(0, 80)}..."`);
-  });
-
-  it('should show +N more when highlights exceed PREVIEW_COUNT (3)', async () => {
-    const highlights = Array.from({ length: 5 }, (_, i) => ({
-      id: String(i),
-      text: `Highlight ${i}`,
-      color: 'yellow',
-    }));
-    await initModule({
-      'highlights_https://d.com/p': { highlights, updatedAt: 1000 },
-    });
-
-    await clickUnsyncedTab();
-
-    const remaining = document.querySelector('.page-card-remaining');
-    expect(remaining.textContent).toContain('還有 2 筆');
-  });
-
-  it('should show load-more button when unsynced pages exceed 10', async () => {
-    const storageData = {};
-    for (let i = 0; i < 15; i++) {
-      storageData[`highlights_https://example.com/page${i}`] = {
-        highlights: [{ id: '1', text: 'x', color: 'yellow' }],
-        updatedAt: i,
-      };
-    }
-    await initModule(storageData);
-    await clickUnsyncedTab();
-
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    expect(cards).toHaveLength(10); // 第一批只顯示 10 張
-    expect(document.querySelector('#load-more-btn').style.display).not.toBe('none');
-  });
-
-  it('should load more cards on load-more click', async () => {
-    const storageData = {};
-    for (let i = 0; i < 15; i++) {
-      storageData[`highlights_https://example.com/page${i}`] = {
-        highlights: [{ id: '1', text: 'x', color: 'yellow' }],
-        updatedAt: i,
-      };
-    }
-    await initModule(storageData);
-    await clickUnsyncedTab();
-
-    // 確認點擊前只有 10 張
-    const container = document.querySelector('#unsynced-view');
-    expect(container.querySelectorAll('.page-card')).toHaveLength(10);
-
-    // 點擊「載入更多」
-    document.querySelector('#load-more-btn').click();
-    for (let i = 0; i < 5; i++) {
-      await Promise.resolve();
-    }
-
-    const cardsAfter = document.querySelectorAll('#unsynced-view .page-card');
-    // 點擊後應比原來多（至少 > 10）
-    expect(cardsAfter.length).toBeGreaterThan(10);
-    expect(document.querySelector('#load-more-btn').style.display).toBe('none');
-  });
-
-  it('should show safe fallback UI when renderUnsyncedView fails after tab switch', async () => {
-    await initModule({});
-
-    const error = new Error('unsynced load failed');
-    chrome.storage.local.get.mockImplementation(async key => {
-      if (key === null) {
-        throw error;
-      }
-      return {};
-    });
-
-    await clickUnsyncedTab();
-
-    expect(Logger.error).toHaveBeenCalledWith(
-      '[SidePanel] renderUnsyncedView failed after tab switch',
-      expect.objectContaining({
-        action: 'renderUnsyncedView',
-        result: 'failure',
-        error,
-      })
-    );
-    expect(document.querySelector('#unsynced-view').textContent).toContain('載入標註失敗');
-    expect(document.querySelector('#unsynced-toolbar').style.display).toBe('none');
-    expect(document.querySelector('#load-more-btn').style.display).toBe('none');
-    expect(document.querySelector('#unsynced-badge').textContent).toBe('');
-  });
-
-  it('should ignore stale unsynced render results after switching back to current quickly', async () => {
-    await initModule({});
-
-    const deferredUnsyncedLoad = createDeferred();
-    chrome.storage.local.get.mockImplementation(async key => {
-      if (key === null) {
-        return deferredUnsyncedLoad.promise;
-      }
-      return {};
-    });
-
-    const unsyncedTab = document.querySelector('[data-view="unsynced"]');
-    const currentTab = document.querySelector('[data-view="current"]');
-
-    unsyncedTab.click();
-    await Promise.resolve();
-    currentTab.click();
-    await Promise.resolve();
-
-    const storageData = {};
-    for (let i = 0; i < 15; i++) {
-      storageData[`highlights_https://example.com/page${i}`] = {
-        highlights: [{ id: '1', text: `x${i}`, color: 'yellow' }],
-        updatedAt: i,
-      };
-    }
-    deferredUnsyncedLoad.resolve(storageData);
-    for (let i = 0; i < 10; i++) {
-      await Promise.resolve();
-    }
-
-    expect(currentTab.classList.contains('active')).toBe(true);
-    expect(unsyncedTab.classList.contains('active')).toBe(false);
-    expect(document.querySelector('#unsynced-toolbar').style.display).toBe('none');
-    expect(document.querySelector('#load-more-btn').style.display).toBe('none');
-  });
-
-  it('should log warning when opening unsynced page fails', async () => {
-    await initModule({
-      'highlights_https://example.com/p': {
-        highlights: [{ id: '1', text: 'open fail', color: 'yellow' }],
-        updatedAt: 1000,
-      },
-    });
-    await clickUnsyncedTab();
-
-    const error = new Error('open failed');
-    chrome.tabs.create.mockRejectedValueOnce(error);
-
-    document.querySelector('.page-open-button').click();
-    await Promise.resolve();
-
-    expect(Logger.warn).toHaveBeenCalledWith('[SidePanel] Failed to open unsynced page tab', {
-      action: 'openUnsyncedPageTab',
-      result: 'failure',
-      error,
-      url: sanitizeUrlForLogging('https://example.com/p'),
-    });
-  });
-
-  it('badge should show correct unsynced count on init', async () => {
-    await initModule({
-      'highlights_https://x.com/p': {
-        highlights: [{ id: '1', text: 'x', color: 'yellow' }],
-        updatedAt: 1,
-      },
-      'highlights_https://y.com/p': {
-        highlights: [{ id: '2', text: 'y', color: 'blue' }],
-        updatedAt: 2,
-      },
-    });
-
-    // badge 在初始化時就應更新
-    const badge = document.querySelector('#unsynced-badge');
-    expect(badge.textContent).toBe('2');
-  });
-
-  it('should log and clear badge when refreshUnsyncedBadge fails during init', async () => {
-    const error = new Error('init badge failed');
-
-    await initModule(async key => {
-      if (key === null) {
-        throw error;
-      }
-      return {};
-    });
-
-    expect(Logger.error).toHaveBeenCalledWith(
-      '[SidePanel] refreshUnsyncedBadge failed during init',
-      expect.objectContaining({
-        action: 'refreshUnsyncedBadge',
-        result: 'failure',
-        error,
-      })
-    );
-    expect(document.querySelector('#unsynced-badge').textContent).toBe('');
-  });
-
-  it('should show empty message when all highlights are synced', async () => {
-    await initModule({
-      'highlights_https://synced.com/p': {
-        highlights: [{ id: '1', text: 'synced', color: 'yellow' }],
-        updatedAt: 1000,
-      },
-      'saved_https://synced.com/p': { notionPageId: 'notion-page-id' },
-    });
-
-    await clickUnsyncedTab();
-
-    const unsyncedView = document.querySelector('#unsynced-view');
-    expect(unsyncedView.textContent).toContain('已全部同步');
-  });
-
-  it('should skip root urls inside getUnsyncedPages (page_* format)', async () => {
-    await initModule({
-      'page_https://example.com/': { highlights: [{ id: '1', text: 'root' }] },
-      'page_https://example.org/': { highlights: [{ id: '2', text: 'another root' }] },
-    });
-    await clickUnsyncedTab();
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    expect(cards).toHaveLength(0); // root urls are skipped
-  });
-
-  it('should skip empty highlights inside getUnsyncedPages (page_* format)', async () => {
-    await initModule({
-      'page_https://example.com/p': { highlights: [] },
-      'page_https://example.com/q': { highlights: null },
-    });
-    await clickUnsyncedTab();
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    expect(cards).toHaveLength(0);
-  });
-
-  it('should include valid page_* items without notion data inside getUnsyncedPages', async () => {
-    await initModule({
-      'page_https://example.com/p': {
-        highlights: [{ id: '1', text: 'valid' }],
-        metadata: { title: 'example.com/p' },
-      },
-    });
-    await clickUnsyncedTab();
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    expect(cards).toHaveLength(1);
-    expect(cards[0].querySelector('.page-title').textContent).toContain('example.com/p');
-  });
-
-  it('should skip root urls inside getUnsyncedPages (highlights_* format)', async () => {
-    await initModule({
-      'highlights_https://example.com/': { highlights: [{ id: '1', text: 'root legacy' }] },
-    });
-    await clickUnsyncedTab();
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    expect(cards).toHaveLength(0);
-  });
-
-  it('should skip empty highlights inside getUnsyncedPages (highlights_* format)', async () => {
-    await initModule({
-      'highlights_https://example.com/p': { highlights: [] },
-      'highlights_https://example.com/q': [],
-    });
-    await clickUnsyncedTab();
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    expect(cards).toHaveLength(0);
-  });
-
-  // === Step 3: Canonical ownership tests（2026-05-03 completion plan） ===
-
-  it('[CANONICAL] page_<stable> 與 page_<original> 並存時應只顯示一張卡片', async () => {
-    await initModule({
-      // alias: original → stable
-      'url_alias:https://example.com/article?ref=x': 'https://example.com/article',
-      // canonical（stable）entry：應被選為 owner
-      'page_https://example.com/article': {
-        notion: null,
-        highlights: [{ id: 'h-stable', text: 'stable text', color: 'yellow' }],
-        metadata: { lastUpdated: 2000, title: 'Stable Title' },
-      },
-      // 殘留的 original entry：alias 命中後應併入同一 canonical group
-      'page_https://example.com/article?ref=x': {
-        notion: null,
-        highlights: [{ id: 'h-orphan', text: 'orphan', color: 'blue' }],
-        metadata: { lastUpdated: 1000, title: 'Orphan Title' },
-      },
-    });
-
-    await clickUnsyncedTab();
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    expect(cards).toHaveLength(1);
-    // owner 應為 page_<stable>，title 對應 stable entry
-    expect(cards[0].querySelector('.page-title').textContent).toContain('Stable Title');
-  });
-
-  it('[CANONICAL] alias 指向空 stable key 時應回退顯示 original page', async () => {
-    await initModule({
-      // alias 存在但 page_<stable> 不存在
-      'url_alias:https://example.com/post?utm=x': 'https://example.com/post',
-      // 只剩 page_<original>
-      'page_https://example.com/post?utm=x': {
-        notion: null,
-        highlights: [{ id: 'h-orig', text: 'original content', color: 'yellow' }],
-        metadata: { lastUpdated: 5000, title: 'Original Page' },
-      },
-    });
-
-    await clickUnsyncedTab();
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    // 不應因為 alias 指向空 stable 而消失；canonical group 仍有 page_<original> 作為 fallback owner
-    expect(cards).toHaveLength(1);
-    expect(cards[0].querySelector('.page-title').textContent).toContain('Original Page');
-  });
-
-  it('[CANONICAL] page_<stable> + highlights_<original> 並存時應依 owner 優先序選 page_<stable>', async () => {
-    await initModule({
-      'url_alias:https://example.com/news?from=fb': 'https://example.com/news',
-      'page_https://example.com/news': {
-        notion: null,
-        highlights: [{ id: 'h1', text: 'canonical', color: 'yellow' }],
-        metadata: { lastUpdated: 9000, title: 'Canonical News' },
-      },
-      'highlights_https://example.com/news?from=fb': [
-        { id: 'h2', text: 'legacy orphan', color: 'blue' },
-      ],
-    });
-
-    await clickUnsyncedTab();
-    const cards = document.querySelectorAll('#unsynced-view .page-card');
-    expect(cards).toHaveLength(1);
-    expect(cards[0].querySelector('.page-title').textContent).toContain('Canonical News');
-  });
-
-  describe('deleteUnsyncedPage', () => {
-    it('should remove the page from storage, cache, and DOM', async () => {
-      const mockKey = 'highlights_https://example.com/delete-test';
-      const mockUrl = 'https://example.com/delete-test';
-      const mockHl = { text: 'delete me', color: 'yellow', id: '1' };
-
-      // 使用 initModule 正確初始化 sidepanel 的 DOM 與內部資源
-      await initModule({
-        [mockKey]: [mockHl],
-      });
-
-      await clickUnsyncedTab();
-
-      const card = document.querySelector('.page-card');
-      expect(card).toBeTruthy();
-
-      const deleteBtn = card.querySelector('.page-delete-button');
-
-      // 模擬 Storage 剛好被清空的狀態（因為只有一台），使後續 getUnsyncedPages 回傳空陣列
-      chrome.storage.local.get.mockResolvedValue({});
-      chrome.runtime.sendMessage.mockResolvedValue({ success: true });
-      chrome.tabs.query.mockResolvedValue([{ id: 777, url: mockUrl }]);
-
-      // Action: 點擊刪除
-      deleteBtn.click();
-      await jest.runAllTimersAsync();
-
-      // Assert storage
-      expect(chrome.storage.local.remove).toHaveBeenCalledWith(mockKey);
-      expect(chrome.runtime.sendMessage).toHaveBeenCalledWith({
-        action: RUNTIME_ACTIONS.CLEAR_HIGHLIGHTS,
-        url: mockUrl,
-        tabId: 777,
-      });
-      expect(chrome.tabs.sendMessage).not.toHaveBeenCalledWith(
-        777,
-        expect.objectContaining({
-          action: RUNTIME_ACTIONS.REMOVE_HIGHLIGHT_DOM,
-        })
-      );
-
-      // Get the current unsynced count label
-      const countLabel = document.querySelector('#unsynced-count-label');
-      expect(countLabel.textContent).toBe('0 個頁面');
-
-      // Trigger CSS animation end to remove card
-      const animationEndEvent = new Event('animationend');
-      card.dispatchEvent(animationEndEvent);
-
-      expect(document.querySelector('.page-card')).toBeNull();
-      expect(document.querySelector('.unsynced-empty')).toBeTruthy();
-    });
-
-    it('[CANONICAL] 從 canonical owner 刪除時 MUST 清除 reverse alias 對應的 page_<original> 與 highlights_<original>', async () => {
-      // Follow-up plan 2026-05-03-highlight-canonical-lock-and-delete-all-cleanup-followup §2：
-      // 當 owner 為 page_<stable> 但 storage 中仍存在 page_<original> / highlights_<original>
-      // （aliasMap 紀錄 <original> → <stable>），刪除應透過 reverse alias 收集並清掉同 canonical group 全體。
-      const stableUrl = 'https://example.com/canonical-target';
-      const originalUrl = 'https://example.com/canonical-target?ref=x';
-      const stablePageKey = `page_${stableUrl}`;
-      const originalPageKey = `page_${originalUrl}`;
-      const originalLegacyKey = `highlights_${originalUrl}`;
-
-      await initModule({
-        [`url_alias:${originalUrl}`]: stableUrl,
-        [stablePageKey]: {
-          notion: null,
-          highlights: [{ id: 'h-stable', text: 'canonical', color: 'yellow' }],
-          metadata: { lastUpdated: 5000, title: 'Canonical Target' },
-        },
-        [originalPageKey]: {
-          notion: null,
-          highlights: [{ id: 'h-orphan', text: 'orphan', color: 'blue' }],
-          metadata: { lastUpdated: 4000, title: 'Orphan Page' },
-        },
-        [originalLegacyKey]: [{ id: 'h-legacy', text: 'legacy', color: 'green' }],
-      });
-
-      await clickUnsyncedTab();
-
-      const card = document.querySelector('.page-card');
-      expect(card).toBeTruthy();
-      expect(card.querySelector('.page-title').textContent).toContain('Canonical Target');
-
-      const deleteBtn = card.querySelector('.page-delete-button');
-      deleteBtn.click();
-      await jest.runAllTimersAsync();
-
-      // chrome.storage.local.remove MUST 被呼叫至少一次,且 keys 涵蓋同 canonical group 全部 member
-      // （listener 可能因 jest.isolateModules 多次重綁,但每次參數應一致;只需任一 call 滿足即可）
-      expect(chrome.storage.local.remove).toHaveBeenCalledWith(
-        expect.arrayContaining([stablePageKey, originalPageKey, originalLegacyKey])
-      );
-    });
-  });
-
-  describe('deleteAllUnsyncedPages', () => {
-    it('should remove all unsynced pages when toolbar Clear All is clicked', async () => {
-      // Setup 2 pages
-      const key1 = 'highlights_https://example.com/p1';
-      const key2 = 'highlights_https://example.com/p2';
-
-      await initModule({
-        [key1]: [{ text: 'hl 1', color: 'yellow', id: '1' }],
-        [key2]: [{ text: 'hl 2', color: 'blue', id: '2' }],
-      });
-
-      await clickUnsyncedTab();
-
-      expect(document.querySelectorAll('.page-card')).toHaveLength(2);
-
-      // 模擬 Storage 已被清空，這樣 await getUnsyncedPages() 才會拿到 0
-      chrome.storage.local.get.mockResolvedValue({});
-
-      // Action: 點擊全部清除
-      const clearBtn = document.querySelector('#clear-all-btn');
-      clearBtn.click();
-
-      await jest.runAllTimersAsync();
-
-      // Assert storage
-      expect(chrome.storage.local.remove).toHaveBeenCalledWith([key1, key2]);
-
-      // Assert DOM
-      expect(document.querySelector('#unsynced-toolbar').style.display).toBe('none');
-      expect(document.querySelector('.unsynced-empty')).toBeTruthy();
-      expect(document.querySelector('#unsynced-badge').textContent).toBe('');
-    });
-
-    it('should do nothing if cachedUnsyncedPages is empty', async () => {
-      await initModule({});
-      await clickUnsyncedTab();
-
-      const clearBtn = document.querySelector('#clear-all-btn');
-      chrome.storage.local.remove.mockClear();
-
-      if (clearBtn && clearBtn.style.display !== 'none') {
-        clearBtn.click();
-      }
-      expect(chrome.storage.local.remove).not.toHaveBeenCalled();
-    });
-
-    it('[CANONICAL] deleteAllUnsyncedPages MUST 清除每張卡片同 canonical group 的全部 keys', async () => {
-      // Follow-up plan 2026-05-03-highlight-canonical-lock-and-delete-all-cleanup-followup §3：
-      // clear-all 路徑必須與 single-delete 一致,使用 canonical-aware 的 cleanup helper,
-      // 不能只刪 entry.storageKey 而漏掉同 canonical group 的其他 member。
-      const stable1 = 'https://example.com/article-one';
-      const original1 = 'https://example.com/article-one?utm=fb';
-      const stable2 = 'https://example.com/article-two';
-
-      await initModule({
-        // Page 1: alias 命中,canonical owner 為 page_<stable1>,但 page_<original1> 殘留
-        [`url_alias:${original1}`]: stable1,
-        [`page_${stable1}`]: {
-          notion: null,
-          highlights: [{ id: 'a-1', text: 'a', color: 'yellow' }],
-          metadata: { lastUpdated: 1000, title: 'Article One' },
-        },
-        [`page_${original1}`]: {
-          notion: null,
-          highlights: [{ id: 'a-orphan', text: 'orphan', color: 'blue' }],
-          metadata: { lastUpdated: 900, title: 'Orphan One' },
-        },
-        // Page 2: 無 alias,canonical owner 直接是 page_<stable2>
-        [`page_${stable2}`]: {
-          notion: null,
-          highlights: [{ id: 'b-1', text: 'b', color: 'yellow' }],
-          metadata: { lastUpdated: 2000, title: 'Article Two' },
-        },
-      });
-
-      await clickUnsyncedTab();
-      expect(document.querySelectorAll('.page-card')).toHaveLength(2);
-
-      // 隔離本次 click 觸發的 remove 呼叫,排除 init / render 階段的雜訊
-      chrome.storage.local.remove.mockClear();
-
-      const clearBtn = document.querySelector('#clear-all-btn');
-      clearBtn.click();
-      await jest.runAllTimersAsync();
-
-      // 至少要有一次 remove call 涵蓋所有 canonical group keys（含 page 1 的 alias original 殘留）
-      expect(chrome.storage.local.remove).toHaveBeenCalledWith(
-        expect.arrayContaining([`page_${stable1}`, `page_${original1}`, `page_${stable2}`])
-      );
-    });
-  });
-});
-
-describe('Required DOM contract', () => {
-  beforeEach(() => {
-    jest.resetModules();
-    jest.clearAllMocks();
-    jest.useFakeTimers();
-
-    document.body.innerHTML = `
-      <div id="loading-state" style="display:none">Loading...</div>
-      <div id="empty-state" style="display:none">
-        <p>Empty</p>
-        <div class="subtitle">Subtitle</div>
-      </div>
-      <div id="highlights-list" style="display:none"></div>
-      <button id="sync-button"></button>
-      <button id="open-notion-button"></button>
-      <div id="status-message"></div>
-      <div id="unsynced-view" style="display:none"></div>
-      <div id="unsynced-toolbar" style="display:none">
-        <span id="unsynced-count-label"></span>
-        <button id="clear-all-btn"></button>
-      </div>
-      <button id="load-more-btn" style="display:none"></button>
-      <span id="unsynced-badge"></span>
-      <button class="view-tab active" data-view="current">Current</button>
-      <button class="view-tab" data-view="unsynced">Pending</button>
-      <template id="highlight-card-template">
-        <div class="highlight-card">
-          <div class="highlight-color-indicator"></div>
-          <p class="highlight-text"></p>
-          <button class="delete-button"></button>
-        </div>
-      </template>
-      <template id="page-card-template">
-        <div class="page-card">
-          <div class="page-title"></div>
-          <div class="page-meta"></div>
-          <div class="page-card-previews"></div>
-          <div class="page-card-remaining"></div>
-          <button class="page-open-button"></button>
-          <button class="page-delete-button"></button>
-        </div>
-      </template>
-    `;
-
-    chrome.tabs.query.mockResolvedValue([{ id: 101, url: 'https://example.com' }]);
-    chrome.tabs.get.mockResolvedValue({ id: 102, url: 'https://example.org' });
-    chrome.tabs.create.mockResolvedValue({ id: 103, url: 'https://opened.example' });
-    chrome.tabs.sendMessage.mockResolvedValue({ stableUrl: 'https://example.js/stable' });
-    chrome.storage.local.get.mockResolvedValue({});
-  });
-
-  afterEach(() => {
-    jest.useRealTimers();
-  });
-
-  it('缺少開始標註按鈕時應立即失敗，避免以不完整 DOM 啟動 sidepanel', async () => {
-    const originalAddEventListener = document.addEventListener.bind(document);
-    let domContentLoadedHandler;
-
-    const addEventListenerSpy = jest
-      .spyOn(document, 'addEventListener')
-      .mockImplementation((eventName, listener, options) => {
-        if (eventName === 'DOMContentLoaded') {
-          domContentLoadedHandler = listener;
-          return;
-        }
-        return originalAddEventListener(eventName, listener, options);
-      });
-
-    try {
-      jest.isolateModules(() => {
-        require('../../../pages/sidepanel/sidepanel.js');
-      });
-
-      expect(typeof domContentLoadedHandler).toBe('function');
-      await expect(domContentLoadedHandler()).rejects.toThrow(
-        '[SidePanel] 缺少必要的 DOM 元素：startHighlightButton'
-      );
-
-      expect(chrome.tabs.onActivated.addListener).not.toHaveBeenCalled();
-    } finally {
-      addEventListenerSpy.mockRestore();
-    }
   });
 });
